+++ types.cpp
@@ -12,7 +12,7 @@ Coords create_coords(double x, double y) {
   return { x, y };
 }
 
-uint8_t coords_distance(Coords* a, Coords* b) {
+uint32_t coords_distance(Coords* a, Coords* b) {
   return sqrt(sq(a->x - b->x) + sq(a->y - b->y)) * DISTANCE_MULTIPLIER;
 }
