+++ types.h
@@ -28,7 +28,6 @@ UID create_uid(EType type, uint8_t x, uint8_t y);
 EType uid_get_type(UID uid);
 
 Coords create_coords(double x, double y);
-uint8_t coords_distance(Coords* a, Coords* b);
+uint32_t coords_distance(Coords* a, Coords* b);
 
 #endif
-
